L3 layout: _groupBuf int* → ushort* (58KB → 29KB; perf flat)
The _groupBuf scratch buffer stores node IDs only. NodeCount is ~15K
which fits ushort (max 65,535). Switched to ushort* for the 29KB saving.

Bit-identical (checksum 0x933ABE7915AC18BE on full_palette.nes 50K hc).

A/B bench on top of the prior `_inGroup byte*` commit (e6824e5):
  prior baseline (byte _inGroup only):  best 38,534  avg 38,018
  with ushort _groupBuf added:           best 38,591  avg 37,874
                                         → +0.15% best, -0.4% avg
                                         → essentially noise

Why no measurable gain (despite the byte _inGroup change giving +3%):
_groupBuf is accessed sequentially (write _groupCount++, read [i] in
the writeback loop), so the cache prefetcher handles it well. Average
group size is 1.4 nodes (per chip-diag), so the array footprint per
walk is tiny regardless of element width.

Kept anyway: memory savings are real (29KB), the (ushort)nn cast is
free on x64 (16-bit store + zero-extend on load are single µops), and
the smaller footprint may help under L2 pressure from future patches.
But the L3-layout series RoI ranking is now clearer:
  _inGroup byte:    +3% (real — random-access bitmap, L1d pressure)
  _groupBuf ushort: 0% (sequential access already cache-friendly)

Next candidate (TransistorList int→short) hits random reads in the
BFS inner loop — likely real win.

+++ src/AprVisual/Sim/WireCore.Group.cs
@@ -27,7 +27,7 @@ namespace AprVisual.Sim
         // Scratch state for the current addNodeToGroup walk (single-threaded).
         private static NodeFlags _groupFlags;
         private static int _groupCount;       // number of nodes currently in _groupBuf
-        private static int* _groupBuf;        // node ids in the current group (alloc'd in Reset, sized NodeCount)
+        private static ushort* _groupBuf;     // node ids in the current group (alloc'd in Reset, sized NodeCount). ushort* (was int*) — node count <65K so 16 bit suffices, 29KB vs 58KB
         private static byte* _inGroup;        // O(1) dedup flag per node (1 = currently in _groupBuf); cleared each ComputeNodeGroup. byte* (was int*) — 0/1 only, 14KB vs 58KB fits L1d alongside NodeStates
         private static byte _maxState;        // state of the highest-connection node seen
         private static int _maxConnections;
@@ -120,7 +120,7 @@ namespace AprVisual.Sim
             _inGroup[nn] = 1;
 
             ref NodeInfo ns = ref NodeInfos[nn];
-            _groupBuf[_groupCount++] = nn;
+            _groupBuf[_groupCount++] = (ushort)nn;
             if (EnableDeadEndDiag && NodeVisitCount != null) NodeVisitCount[nn]++;
             if (EnableChipDiag)
             {
